Make unit tests pass on TravisCI.
